+++ source/js/main.js
@@ -1,4 +1,0 @@
-// import vendor/_modernizr.js
-// import ../../bower_components/requirejs/require.js
-// import _requireconfig.js
-// import _app.js