+++ source/js/modules/common.js
@@ -1,1 +1,0 @@
-// import ../vendor/plugins/_domReady.js